requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
crispy-bootstrap5==0.7
Django==4.1.7
django-bootstrap4==22.3
django-crispy-forms==2.0
django-filter==22.1
Pillow==9.4.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
